Ditched dual-gain mode
There is a problem where the GPIO LOW state pulls to ~3.5mV rather than
0V as expected due to [supply current] * [resistance between the IC
ground and PCB ground]. Using superposition, this ~3.5 mV sees an
inverting amplifier with gain -Rf/Rs and causes an offset in the
measured value, unfortunately limiting our low-end sensitivity. I
mitigated this by increasing the transimpedance gain and reducing the
voltage gain. With some averaging, we get clean measurements in the
range of interest.